backend/api/urls.py
from django.urls import include, path
from rest_framework import routers
from api.views import (IngredientViewSet, TagViewSet, RecipeViewSet)
from users.views import CustomUserViewSet

app_name = 'api'

router = routers.DefaultRouter()
router.register('ingredients', IngredientViewSet)
router.register('tags', TagViewSet)
router.register('recipes', RecipeViewSet)
router.register('users', CustomUserViewSet)


urlpatterns = [
    path('', include(router.urls)),
    path('', include('djoser.urls')),
    path('auth/', include('djoser.urls.authtoken')),
]
